Clean-up, and moved banner edit code from profile page to editprofilepage
