+++ Frontend/src/Pages/EditProfilePage.jsx
@@ -7,10 +7,16 @@ import { useNavigate } from "react-router-dom";
 import { toast } from "react-hot-toast";
 
 const EditProfilePage = (props) => {
-	const { updateTab, userDetails } = props;
+	const {
+		updateTab,
+		userDetails,
+		updatePage: fetchData,
+		profileOwner,
+	} = props;
 	const User = UseUser();
 	const CustomFetch = useFetch();
 	const Navigate = useNavigate();
+	const [bannerList, setBannerList] = useState([]);
 	const [user, setUser] = useState();
 	const [isLoading, setIsLoading] = useState(true);
 	const [data, setData] = useState({
@@ -25,35 +31,38 @@ const EditProfilePage = (props) => {
 		// profilePicture: "",
 	});
 
+	const fetchUserData = async () => {
+		const { response, data } = await CustomFetch({
+			url: `user/${profileOwner}`,
+			returnResponse: true,
+		});
+
+		if (!response.ok) {
+			Navigate("/notfound");
+		}
+
+		setData(() => {
+			return {
+				username: data.data.username,
+				email: data.data.email,
+				bio: data.data.bio,
+				discord: data.data.discord,
+				instagram: data.data.instagram,
+				mc_username: data.data.mc_username,
+				showMail: data.data.showMail,
+				profilePicture: data.data.profilePicture,
+				banner: data.data.banner.url,
+				bannerId: data.data.banner.id,
+			};
+		});
+
+		setUser(data.data);
+		setIsLoading(false);
+	};
+
 	useEffect(() => {
 		(async () => {
-			const { response, data } = await CustomFetch({
-				url: `user/${User.id}`,
-				returnResponse: true,
-			});
-
-			if (!response.ok) {
-				Navigate("/notfound");
-			}
-			console.log(data.data);
-
-			setData(() => {
-				return {
-					username: data.data.username,
-					email: data.data.email,
-					bio: data.data.bio,
-					discord: data.data.discord,
-					instagram: data.data.instagram,
-					mc_username: data.data.mc_username,
-					showMail: data.data.showMail,
-					profilePicture: data.data.profilePicture,
-					banner: data.data.banner.url,
-					bannerId: data.data.banner.id,
-				};
-			});
-
-			setUser(data.data);
-			setIsLoading(false);
+			await fetchUserData();
 		})();
 	}, []);
 
@@ -92,13 +101,77 @@ const EditProfilePage = (props) => {
 		});
 	};
 
+	useEffect(() => {
+		(async () => {
+			const { response, data } = await CustomFetch({
+				url: "banner",
+				returnResponse: true,
+			});
+
+			if (!response.ok) return;
+			// ! handle error
+			setBannerList(data.data);
+		})();
+	}, []);
+
+	const updateBanner = async (bannerId) => {
+		const Request = CustomFetch({
+			url: `user/${User.id}`,
+			options: {
+				method: "PATCH",
+				body: JSON.stringify({
+					bannerId: bannerId,
+				}),
+			},
+			returnPromise: true,
+		});
+
+		toast.promise(Request, {
+			loading: "Updating banner...",
+			success: (data) => {
+				(async () => {
+					await fetchData();
+					await fetchUserData();
+					// TODO: Getting user's data from the main profile page.
+				})();
+				return `Sucessfully updated banner!`;
+			},
+			error: (err) => {
+				console.log(err);
+				return `An error occured while trying to update your banner!`;
+			},
+		});
+	};
+
 	return (
 		<>
 			{!isLoading && (
 				<div className="bg-white">
 					<div className="h-[500px] grid place-content-center relative">
+						<div className="bg-red-400 w-full h-[200px] fixed m-4 z-50">
+							One
+						</div>
 						<div className="absolute bg-transparent grid place-content-center z-20 top-0 bottom-0 left-0 right-0">
 							<div className="w-[250px] h-[250px] rounded-full hover:cursor-pointer">
+								<div className="absolute top-0 right-0 p-2 flex flex-row gap-1">
+									{/* Need to change the banner list to data pulled from DB - currently saved on FE */}
+									{User.isAuthenticated &&
+										(User.id === user.id ||
+											User.role.isAdmin) &&
+										bannerList?.map((item) => {
+											const { id, url } = item;
+											return (
+												<img
+													key={id}
+													src={url}
+													onClick={() =>
+														updateBanner(id)
+													}
+													className="w-[32px] h-[32px] cursor-pointer border border-gray-500"
+												/>
+											);
+										})}
+								</div>
 								<img
 									src={userDetails.profilePicture}
 									alt=""
